🚚 [variables] Move module `bindings` to `domain`
