+++ src/cutty/entrypoints/cli/create.py
@@ -7,7 +7,7 @@ import click
 from cutty.entrypoints.cli.cookiecutter import extra_context_callback
 from cutty.entrypoints.cli.errors import fatal
 from cutty.services.create import create as service_create
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 
 
 @click.command()
