+++ src/cutty/services/link.py
@@ -12,7 +12,7 @@ from cutty.projects.config import readcookiecutterjson
 from cutty.projects.config import readprojectconfigfile
 from cutty.projects.messages import linkcommitmessage
 from cutty.projects.repository import ProjectRepository
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 
 
 class TemplateNotSpecifiedError(CuttyError):
