+++ src/cutty/services/update.py
@@ -8,7 +8,7 @@ from cutty.projects.config import ProjectConfig
 from cutty.projects.config import readprojectconfigfile
 from cutty.projects.messages import updatecommitmessage
 from cutty.projects.repository import ProjectRepository
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 
 
 def update(
