+++ src/cutty/templates/adapters/jinja.py
@@ -14,7 +14,7 @@ import jinja2
 from cutty.filesystems.domain.path import Path
 from cutty.templates.domain.render import GenericRenderer
 from cutty.util.reraise import reraise
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 
 
 @dataclass
