+++ src/cutty/templates/domain/bindvariables.py
@@ -4,7 +4,7 @@ from collections.abc import Sequence
 from cutty.templates.domain.render import Renderer
 from cutty.variables.binders import Binder
 from cutty.variables.binders import override
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 from cutty.variables.domain.variables import Variable
