+++ src/cutty/templates/domain/render.py
@@ -13,7 +13,7 @@ from cutty.filestorage.domain.files import RegularFile
 from cutty.filestorage.domain.files import SymbolicLink
 from cutty.filesystems.domain.path import Path
 from cutty.filesystems.domain.purepath import PurePath
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 from cutty.variables.domain.variables import GenericVariable
 from cutty.variables.domain.variables import Variable
