+++ src/cutty/templates/domain/renderfiles.py
@@ -7,7 +7,7 @@ from cutty.filestorage.domain.files import File
 from cutty.filestorage.domain.files import loadfile
 from cutty.filesystems.domain.path import Path
 from cutty.templates.domain.render import Renderer
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 
 
 def renderfiles(
