+++ src/cutty/variables/prompts.py
@@ -15,7 +15,7 @@ from questionary import Validator
 from cutty.variables.binders import bind
 from cutty.variables.binders import binddefault
 from cutty.variables.binders import Binder
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 from cutty.variables.domain.variables import Variable
