+++ tests/unit/templates/domain/test_bindings.py
@@ -3,7 +3,7 @@ from typing import Any
 
 import pytest
 
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 
 
 @pytest.mark.parametrize(
