+++ tests/unit/templates/domain/test_bindvariables.py
@@ -5,7 +5,7 @@ from cutty.templates.domain.bindvariables import renderbind
 from cutty.templates.domain.render import Renderer
 from cutty.variables.binders import binddefault
 from cutty.variables.binders import override
-from cutty.variables.bindings import Binding
+from cutty.variables.domain.bindings import Binding
 from cutty.variables.domain.variables import GenericVariable
