+++ tests/unit/variables/test_prompts.py
@@ -9,7 +9,7 @@ from prompt_toolkit.input.defaults import create_pipe_input
 from prompt_toolkit.output import DummyOutput
 
 from cutty.variables.binders import Binder
-from cutty.variables.bindings import GenericBinding
+from cutty.variables.domain.bindings import GenericBinding
 from cutty.variables.domain.variables import GenericVariable
 from cutty.variables.prompts import createprompt
 from tests.util.keys import BACKSPACE
